requirements.txt
notion-client==2.2.1
openai==1.42.0
nltk==3.8.1
matplotlib==3.8.2
pandas==2.1.4
python-dotenv==1.0.1
requests==2.32.3
